book-service/requirements.txt
Django==6.0.3
djangorestframework==3.16.1
requests>=2.28.0
psycopg2-binary>=2.9.0
drf-spectacular==0.28.0
